[feature/메인 강남구지역 매장리스트조회 구현]
