+++ src/main/java/com/kkosoonnae/jpa/projection/MainStoresListviewProjection.java
@@ -1,0 +1,9 @@
+package com.kkosoonnae.jpa.projection;
+
+public record MainStoresListviewProjection(
+        Integer storeNo,
+        String storeName,
+        String roadAddress,
+        Double scope
+) {
+}
