+++ src/main/java/com/kkosoonnae/search/dto/MainStoreListViewResponseDto.java
@@ -1,35 +1,33 @@
 package com.kkosoonnae.search.dto;
 
-import com.kkosoonnae.jpa.projection.StoreListViewProjection;
+import com.kkosoonnae.jpa.projection.MainStoresListviewProjection;
 import io.swagger.v3.oas.annotations.media.Schema;
-import lombok.*;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
 
 @Getter
-@Setter
 @AllArgsConstructor
 @NoArgsConstructor
 @Builder
-public class StoreListViewResponseDto {
-
+public class MainStoreListViewResponseDto {
     @Schema(description = "매장 번호")
     private Integer storeNo;
     @Schema(description = "매장 이름")
     private String storeName;
     @Schema(description = "매장 주소")
     private String roadAddress;
-    @Schema(description = "매장 이미지")
-    private String img;
     @Schema(description = "평균 별점")
     private Double averageScope;
-   // private Long totalLikeStore;
 
-    public static StoreListViewResponseDto projectToDto(StoreListViewProjection projection) {
-        return StoreListViewResponseDto.builder()
+    public static MainStoreListViewResponseDto mainStoresDto (MainStoresListviewProjection projection) {
+        return MainStoreListViewResponseDto.builder()
                 .storeNo(projection.storeNo())
                 .storeName(projection.storeName())
                 .roadAddress(projection.roadAddress())
-                .img(projection.img())
                 .averageScope(projection.scope())
                 .build();
     }
+
 }
